# packages/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 模序前后端协作实验
  version: 0.1.0
paths:
  /api/auth/request-code:
    post:
      summary: 为初版注册或登录流程生成页面验证码
      responses:
        '200': { description: 页面验证码已生成，不发送短信 }
        '422': { description: 手机号格式不正确 }
  /api/auth/register:
    post:
      summary: 使用手机号、验证码和密码注册
      responses:
        '200': { description: 注册成功并返回会话令牌 }
        '422': { description: 验证码或输入无效 }
  /api/auth/login:
    post:
      summary: 使用手机号账号与密码或页面验证码登录
      responses:
        '200': { description: 登录成功并返回会话令牌 }
        '401': { description: 手机号或密码错误 }
  /api/chat:
    post:
      summary: 与云朵助手“Mochi”进行流式对话
      responses:
        '200': { description: UTF-8 流式文本回复 }
        '503': { description: 服务端尚未配置模型 }
  /api/v1/conversations/{conversationId}/signals:
    post:
      summary: 提交设备端生成的短时表情辅助信号
      description: 不接收原始照片、视频、人脸关键点或身份特征；需要用户事先开启表情辅助权限。
      parameters:
        - in: path
          name: conversationId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [label, intensity, confidence]
              properties:
                label: { type: string, enum: [calm, tired, tense, positive, uncertain] }
                intensity: { type: number, minimum: 0, maximum: 1 }
                confidence: { type: number, minimum: 0, maximum: 1 }
      responses:
        '200': { description: 信号已保存至内存并将在十分钟后失效 }
        '403': { description: 用户尚未开启表情辅助权限 }
        '422': { description: 信号格式无效 }
    delete:
      summary: 立即撤回当前对话的临时表情辅助信号
      parameters:
        - in: path
          name: conversationId
          required: true
          schema: { type: string }
      responses:
        '200': { description: 临时信号已删除 }
  /api/comfort:
    post:
      summary: 根据用户记录返回陪伴草稿和小行动
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [event]
              properties:
                event: { type: string, minLength: 1 }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                required: [reply, actions, contractVersion]
                properties:
                  reply: { type: string }
                  actions: { type: array, items: { type: string } }
                  contractVersion: { type: string }
        '422': { description: 输入为空或不符合契约 }